Added feature to include results generation
