+++ pom.xml
@@ -4,7 +4,7 @@
 
   <groupId>com.checkmarx.ast</groupId>
   <artifactId>ast-cli-java-wrapper</artifactId>
-  <version>1.0.7</version>
+  <version>1.0.9</version>
   <packaging>jar</packaging>
 
   <dependencies>
