--- src/main/java/com/checkmarx/ast/CxCommandOutput.java
@@ -1,15 +1,0 @@
-package com.checkmarx.ast;
-
-
-import lombok.*;
-
-
-import java.util.List;
-
-@Getter
-@Setter
-public class CxCommandOutput {
-    private int exitCode;
-    private List<CxScan> scanObjectList;
-
-}
