--- src/main/java/com/checkmarx/ast/CxException.java
@@ -1,7 +1,0 @@
-package com.checkmarx.ast;
-
-public class CxException extends RuntimeException{
-    public CxException(String errorMessage) {
-        super(errorMessage);
-    }
-}
