--- src/main/java/com/checkmarx/ast/CxParamType.java
@@ -1,5 +1,0 @@
-package com.checkmarx.ast;
-
-public enum CxParamType {
-    S, V, G, PROJECT_NAME, SCAN_TYPES, SAST_PRESET_NAME, FILTER, DIRECTORY, ADDITIONAL_PARAMETERS, AGENT, SOURCES, BRANCH
-}
