+++ src/main/java/com/checkmarx/ast/exceptionhandler/CxException.java
@@ -1,0 +1,7 @@
+package com.checkmarx.ast.exceptionhandler;
+
+public class CxException extends RuntimeException{
+    public CxException(String errorMessage) {
+        super(errorMessage);
+    }
+}
