+++ src/main/java/com/checkmarx/ast/resultshandler/CxResultFormatType.java
@@ -1,0 +1,5 @@
+package com.checkmarx.ast.resultshandler;
+
+public enum CxResultFormatType {
+    JSON, LIST
+}
