+++ src/main/java/com/checkmarx/ast/resultshandler/CxResultType.java
@@ -1,0 +1,5 @@
+package com.checkmarx.ast.resultshandler;
+
+public enum CxResultType {
+    LIST, SUMMARY
+}
