+++ src/main/java/com/checkmarx/ast/scanhandler/CxParamType.java
@@ -1,0 +1,5 @@
+package com.checkmarx.ast.scanhandler;
+
+public enum CxParamType {
+    S, V, G, PROJECT_NAME, SCAN_TYPES, SAST_PRESET_NAME, FILTER, DIRECTORY, ADDITIONAL_PARAMETERS, AGENT, SOURCES, BRANCH
+}
