+++ src/test/java/com/checkmarx/ast/CxAuthTest.java
@@ -8,13 +8,22 @@ import org.junit.runners.JUnit4;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.File;
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
+import com.checkmarx.ast.resultshandler.CxResultFormatType;
+import com.checkmarx.ast.resultshandler.CxResultOutput;
+import com.checkmarx.ast.resultshandler.CxResultType;
+import com.checkmarx.ast.scanhandler.CxAuth;
+import com.checkmarx.ast.scanhandler.CxCommandOutput;
+import com.checkmarx.ast.scanhandler.CxParamType;
+import com.checkmarx.ast.scanhandler.CxScanConfig;
+
 import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @RunWith(JUnit4.class)
 public class CxAuthTest {
@@ -78,7 +87,8 @@ public class CxAuthTest {
         params.put(CxParamType.BRANCH, "test");
 
         CxCommandOutput scanResult = auth.cxScanCreate(params);
-        assertTrue(auth.cxScanShow(scanResult.getScanObjectList().get(0).getID()).getScanObjectList().get(0).getStatus().equalsIgnoreCase(COMPLETED));
+        String status = auth.cxScanShow(scanResult.getScanObjectList().get(0).getID()).getScanObjectList().get(0).getStatus();
+        assertTrue(status.equalsIgnoreCase(COMPLETED));
     }
 
     @Test
@@ -87,8 +97,8 @@ public class CxAuthTest {
         params.put(CxParamType.SAST_PRESET_NAME, "Checkmarx Default Jay");
 
         CxCommandOutput scanResult = auth.cxScanCreate(params);
-
-        assertTrue(auth.cxScanShow(scanResult.getScanObjectList().get(0).getID()).getScanObjectList().get(0).getStatus().equalsIgnoreCase(FAILED));
+        String status = auth.cxScanShow(scanResult.getScanObjectList().get(0).getID()).getScanObjectList().get(0).getStatus();
+        assertTrue(status.equalsIgnoreCase(FAILED));
     }
 
 
@@ -100,5 +110,22 @@ public class CxAuthTest {
 
         CxCommandOutput scanResult = auth.cxScanCreate(params);
         assertTrue(auth.cxScanShow(scanResult.getScanObjectList().get(0).getID()).getScanObjectList().get(0).getStatus().equalsIgnoreCase(COMPLETED));
+    }
+
+    @Test
+    public void cxGetListResults() throws InterruptedException, IOException {
+        CxCommandOutput scanList = auth.cxAstScanList();
+        String id = scanList.getScanObjectList().get(0).getID();
+        CxResultOutput scanResult = auth.cxGetResults(CxResultType.LIST, id, CxResultFormatType.JSON, null);
+        assertTrue(scanResult.getResults().size() > 0);
+    }
+
+    @Test
+    public void cxGenerateHTMLResults() throws InterruptedException, IOException {
+        CxCommandOutput scanList = auth.cxAstScanList();
+        String id = scanList.getScanObjectList().get(0).getID();
+        String filePath = System.getProperty("user.dir") + "/index.html";
+        auth.cxGetResults(CxResultType.SUMMARY, id, null, filePath);
+        assertTrue(new File(filePath).length() > 0);
     }
 }
